move checks to checks.py
All check-related functions from functions.py now reside in utils/checks.py
